link2symlink: Don't decrement link count on rmdir()
The unlink() and rmdir() calls in Linux got merged into unlinkat()
with AT_REMOVEDIR flag to differentiate them.

Previously link2symlink handled unlink and unlinkat equivalently
and not handled rmdir() at all. However as unlinkat can mean rmdir
we have to check this and not count it as decrement of link count.

Dpkg when removes files first tries to rmdir() and if that fails with
-ENOTDIR then uses unlink()

Fixes termux/proot#18