--input and --output -> --input-target and --output-target
When I was trying to compile in I don't remember which Termux or Arch Linux in Termux. I was getting error of --input and --output option doesn't exist. In the manpage and google suggest --input-target and --output-target.

http://man7.org/linux/man-pages/man1/objcopy.1.html

Although it doesn't have any issue in Debian but I think we should follow what it's in the manpage.