Remove prepare_restart_syscall_after_seccomp function
It originally existed as there was requirement
to call it iff restart_syscall_after_seccomp was called after
but that requirement was inconvenient and was removed some time
ago. The result was that every handler called that function
at beginning regardless of what it was going to do, so we now
perform that preparation always before dispatching to syscall
handler and don't need separate preparation function.